+++ app/src/main/java/de/baumann/browser/objects/CustomRedirectsHelper.java
@@ -1,6 +1,5 @@
 package de.baumann.browser.objects;
 
-import android.content.Context;
 import android.content.SharedPreferences;
 
 import androidx.preference.PreferenceManager;
@@ -10,6 +9,9 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
+import java.util.Objects;
+
+import de.baumann.browser.activity.BrowserActivity;
 
 public class CustomRedirectsHelper {
 
@@ -19,13 +21,13 @@ public class CustomRedirectsHelper {
         ArrayList<CustomRedirect> redirects = new ArrayList<>();
         String redirectsPref = preferences.getString(CUSTOM_REDIRECTS_KEY, "[]");
 
-        if (preferences.getBoolean("sp_youTube_switch", false)) {
+        if (Objects.requireNonNull(preferences.getString("saved_redirect_ok", "no")).equals("no")) {
             redirects.add(new CustomRedirect("m.youtube.com", preferences.getString("sp_youTube_string_domain", "yewtu.be")));
             redirects.add(new CustomRedirect("youtube.com", preferences.getString("sp_youTube_string_domain", "yewtu.be")));
-        }
-
-        if (preferences.getBoolean("sp_twitter_switch", false)) {
             redirects.add(new CustomRedirect("twitter.com", preferences.getString("sp_twitter_string_domain", "nitter.net")));
+            saveRedirects(redirects);
+            preferences.edit().putString("saved_redirect_ok", "yes").apply();
+
         }
 
         JSONArray array = new JSONArray(redirectsPref);
@@ -38,8 +40,8 @@ public class CustomRedirectsHelper {
         return redirects;
     }
 
-    public static void saveRedirects(Context context, ArrayList<CustomRedirect> redirects) throws JSONException {
-        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(context);
+    public static void saveRedirects(ArrayList<CustomRedirect> redirects) throws JSONException {
+        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(BrowserActivity.getAppContext());
         JSONArray array = new JSONArray();
         for (int i = 0; i < redirects.size(); i++) {
             CustomRedirect redirect = redirects.get(i);