+++ app/src/main/java/de/baumann/browser/view/NinjaWebView.java
@@ -660,7 +660,7 @@ public class NinjaWebView extends WebView implements AlbumController {
             stopLoading();
             GridItem item_01 = new GridItem("https://", R.drawable.icon_secure);
             GridItem item_02 = new GridItem( "http://", R.drawable.icon_unsecure);
-            GridItem item_03 = new GridItem( context.getString(R.string.dialog_neverAsk), R.drawable.icon_close);
+            GridItem item_03 = new GridItem( context.getString(R.string.app_never), R.drawable.icon_close);
 
             View dialogView = View.inflate(context, R.layout.dialog_menu, null);
             MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(context);